+++ apps/frontend/src/app/3_organisms/LOCChart/LOCChart.tsx
@@ -262,7 +262,8 @@ export const LOCChart: FC = () => {
       userOpenTroveAbove &&
       userOpenTroveBelow &&
       !loadingUserOpenTroveAbove &&
-      !loadingUserOpenTroveBelow
+      !loadingUserOpenTroveBelow &&
+      userOpenTrove.trove
     ) {
       const { transaction, trove } = userOpenTrove.trove.changes[0];
